try to solve problems in Model.pred function, failed 已经把问题定位到了 model.pred 这个函数的参数传递上， 现在就是在 _execute 这个位置是不会出现任何的data assertion error 但是当 data 的指针被传递给了pred函数时，就有可能出现问题，而且这个问题不是稳定复现的

+++ src/utils.cpp
@@ -21,7 +21,7 @@ double BoxMullerRandom() {
 }
 
 double random01() {
-    return rand() / RAND_MAX;
+    return (double)rand() / RAND_MAX;
 }
 
 void parseArgs(int argc, char *argv[])
